requirements.txt
colorama
flask
flask_httpauth
werkzeug
pyopenssl
termcolor
colored
cprint
zeroconf
IPVersion
ServiceInfo
Zeroconf